pyproject.toml
[project]
name = "rowan_sample_env"
description = "Rowan Sample Environment"
version = "0.0.1"
readme = "README.md"
keywords = []
requires-python = ">=3.13"
dependencies = [
    "stjames",
    "rowan-python",
]

[dependency-groups]
dev = [
    "prek>=0.2",
    "ruff>=0.14",
    "ty>=0.0.11",
]

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = [
    "B",  # bugbear
    "D",  # pydocstyle
    "E",  # pycodestyle errors
    "F",  # pyflakes
    "I",  # isort
    "N",  # pep8-naming conventions
    "W",  # pycodestyle warnings
    "C4", # comprehensions
    "PL", # pylint
    "PT", # flake8-pytest-style
    "PIE", # misc lints
    "PYI", # flake8-pyi
    "TID", # tidy imports
    "TCH", # type-checking imports
    "RUF", # Ruff-specific rules
    "RSE", # flake8-raise
    "ICN001", # unconventional-import-alias
]
ignore = [
    "N806", # Non-lowercase variable in function
    "PLR0911", # Too many returns
    "PLR0912", # Too many branches
    "PLR0913", # Too many arguments to function call
    "PLR0914", # Too many locals
    "PLR0915", # Too many statements
    "PLR1702", # Too many nested-blocks
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401", "F403"]

[tool.ruff.lint.pydocstyle]
convention = "google"
